// src/types/vinho.ts
export interface Vinho {
  avaliations: number; 
  classification: string;
  country: string; 
  discount: number;  
  flag: string;
  image: string;
  name: string; 
  price: number;  
  priceMember:number;  
  priceNonMember: number; 
  rating: number; 
  region: string;
  size: number;  
  sommelierComment: string;
  type: string;
  id: number;
  volume: number;
}